requirements.txt
statsig-python-core
setuptools
opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-api
python-dotenv
pytest
pytest-httpserver
werkzeug